Add .dockerignore file
This commit introduces a .dockerignore file to the project. With this update, unnecessary files and directories (such as .idea, .run, .git, and log files) are excluded from the Docker build context, improving build performance and reducing the image size. Also, sensitive application properties files from the backend are ignored for security purposes.