Add user registration and retrieval services
Three new files are introduced: UserService.java, UserRepository.java, and AppUser.java in the user directory. UserService handles registration and retrieval of user data. UserRepository is an interface with the MongoDB repository that facilitates the operations of the UserService. AppUser is the data model for user information. Changes were also made to the SecurityConfig.java file to accommodate user registration and retrieval services.