Add content creation functionality to ContentController
A new endpoint has been added to ContentController to handle the creation of new Content. The controller now uses userService in addition to contentService, therefore the userService object has been added. A new class called NewContentDTO has been created to send new content data from client to server. The endpoint works by taking in a principal and a request body containing a NewContentDTO from which it extracts content data, associates it to an AppUser, and saves it using createContent method in ContentService.