Add content count to content filter page
A new `infoType` was introduced to return data along with the content information on the content filter page. This addition includes a change in the content service to return `info` and `content` instead of just `content`. A new component, `ContentFilterResultHeader`, was also added to display the number of content returned.