Add tests for ContentCard component
The commit includes adding `@testing-library/user-event` to the package.json dependencies and creating a new test file for the ContentCard component. The tests cover rendering of English and German titles, correct SVG icons based on content type, and the display of a tooltip on icon hover. To facilitate testing, data-testids have been added to the SVG icons in ContentCardTypeIcon component.