Implement soft deletion functionality for Content
A soft delete feature has been added to the Content class which sets the ContentStatus to DELETED instead of completely removing it. This includes creating a new ContentStatus enum, updating the ContentService to perform a soft delete, and providing an endpoint in the ContentController. Also, the UserService has been injected into ContentService for user validations during a soft delete. Lombok has been added as a dependency for code simplification.