Refactor frontend components and improve code readability
This commit involves a major refactor of the frontend components to enhance code clarity and reusability. A new component `ContentPrimaryViewSection` is created for displaying data in the `ContentPrimaryView` file. This is aimed at reducing code repetition and allowing for more modular design. The returned datatype of the `useGithubUserById` function in the `githubService` has also been made explicit.