Update test data structure and make editable prop optional in Scopebox
The test data structure in ContentResultBody and ContentPrimaryView has been updated for simplified usage. A user object was created and used in createdBy field of the content objects. Additionally, status fields have been added to content objects. The editable prop in Scopebox component is now optional, improving the component's flexibility.