Add react-hook-form library and implement in ContentCreationDrawer
The react-hook-form library has been added to the project dependencies. This library was used to refactor the content creation form in the ContentCreationDrawer component. This provides better form handling and validation rules. An axios HTTP post request has also been added to submit the form data to the "api/content" endpoint.