Add test for updating content when it's deleted
This commit adds a new test case 'updateContentByIdTest_whenContentDeleted_thenReturnBadRequest' in the 'ContentControllerTest' class. This test checks if the API returns a bad request when trying to update a content that has been deleted. The test first creates a content, then deletes it and tries to update it, expecting a BadRequest status.