Add user retrieval functionality and refactor code
A `getAllUsers()` method was added to `UserService`. This method facilitates the retrieval of all application users from the repository. Additionally, the `UserController` file was created to enable API calls for user details. On the frontend side, updated the tooltip message in `ContentModalEditTitles`. Also refactored `ContentController` to use `@RequiredArgsConstructor` annotation for cleaner code.