Add GitHub user profile syncing functionality
A GitHub service is added to get user profiles and their synchronization has been implemented in the userService. A GithubMapper class is introduced to map OAuth2User attributes to GitHub profiles. AppUsers are updated to include profile sync time information and UserService's register method is refactored to use the new mapping function.