Add synchronizing Github user profiles feature
The update includes a new function in UserService called 'syncGithubUserProfile' which helps determine whether a Github profile has been updated since the last sync. If updated, this method syncs the new Github profile with the AppUser and updates the sync time. The 'UserController' has also been extended to facilitate post requests for this synchronization process.