Refactor services and update type definitions
The code updates and refactors the services to provide better type safety and consistency. In particular, it modifies the content service to set 'AxiosError' as the type for error and refactors the response structure. Type 'AppUserStatusENUM' is moved and 'GithubUserSyncedType' is exported in userModel. A method is added in githubService to process 'GithubUserSyncedType'. userService is refactored to process user data and handle errors more effectively.