Add StaffTable component and adjust userModel attributes
A new StaffTable component was added to the system that displays information about each user. The model for the user data, userModel, had several attributes adjusted for optional properties and a typo was corrected in the "githubUserProfileSynced" attribute. A console.log was also added in userService to allow for debugging.