Refactor ContentResultHeader component to ResultHeader
This commit renames the ContentResultHeader component to a more reusable name, ResultHeader. The tests associated with the component have also been updated. In addition, usage of the old component in the codebase has been replaced with the new component, and a skeleton state for the staff page has been introduced.