Add 'getPersonById' endpoint and 'PersonNotFoundException'
Implemented a new endpoint to fetch a person by their ID in the PersonController. Also added a new custom exception 'PersonNotFoundException' which is thrown if no person is found with a given ID. The exception is also handled globally. Improvements in test cases to cover these changes were also made.