Add validation for person and content creation
Extended validation for the creation of new 'Person' and 'Content' entities in both frontend and backend. The first name and last name fields of a 'Person' are now limited to a maximum of 50 characters. For both 'Person' and 'Content', an ID must be null when creating a new entity to prevent duplicate entries.