Add people management functionality to frontend
This commit introduces several enhancements focused on people management. It includes a new 'PeoplePage' and related service functions to display a list of people from the API. Also, it adds a 'PersonTable' component for structured presentation of people data. In addition, it modifies some existing components and models to support the new functionality.