Update code for improved data handling and add staff table row
This commit incorporates updates in various places of the code to handle data more effectively and consistently. The significant change was to refactor the table handling for staff members where each staff member now gets their own StaffTableRow component. This leads to cleaner code, allows for better separation of concerns, and introduces the possibility for expanding the functionalities of each row separately in the future. Additionally, to correspond to recent changes, the appUser import was updated to appUser1 across multiple files, and keys were added to React components when mapping over data arrays to prevent warnings.