TEMP: Add LoginModal and auth checks to actions
A new LoginModal component has been added for handling user session expiration. Any action that requires authentication in the NavigationSidebar now checks if the user is logged in before execution. UseLoginModalStore was created for managing the state of the LoginModal, and its implementation was integrated into App.tsx. An interceptor was also added to trigger the LoginModal when a 401 error is received from the server.