Add About page to the frontend
A new page, "About", has been added to the frontend application. With this, a route "/about" has been established to direct users to this page. The Navbar also includes a link to this new page for easy navigation. The About page, apart from providing general information, also serves to give credit to contributors.