Add onClick prop to Logo component
The Logo component now takes an onClick prop to add interactivity. This update also includes the inclusion of a cursor style prop and modification of the internal structure of the Logo component. These changes aim to enhance the functionality and flexibility of the component in different contexts, for example in NavigationSidebar.