Remove logback-test.xml and update backend tokens
The logback-test.xml file is deleted as it's no longer needed. Meanwhile, the BETTERSTACK_TOKEN_BACKEND values in the application properties files have been updated, clearly separating them into production and test configurations. This is a necessary move to segregate testing from production to maintain the integrity of the production data.