Update README with JDK version and credits
The README file has been updated to specify the Java JDK version required for the project. Additionally, a "Credits" section has been added to acknowledge the contributions of specific individuals to the project development and design. A "Special Thanks" section has also been added to extend appreciation to some people who contributed exceptionally to the project.